+++ packages/datadog_flutter_plugin/lib/src/logs/ddlogs_web.dart
@@ -24,7 +24,7 @@ class DdLogsWeb extends DdLogsPlatform {
       clientToken: configuration.clientToken,
       env: configuration.env,
       site: siteStringForSite(configuration.site),
-      proxyUrl: configuration.customEndpoint,
+      proxyUrl: configuration.customLogsEndpoint,
       service: configuration.serviceName,
       version: version,
     ));