+++ packages/datadog_tracking_http_client/example/lib/main.dart
@@ -50,7 +50,7 @@ Future<void> main() async {
     batchSize: BatchSize.small,
     nativeCrashReportEnabled: true,
     firstPartyHosts: firstPartyHosts,
-    customEndpoint: customEndpoint,
+    customLogsEndpoint: customEndpoint,
     loggingConfiguration: LoggingConfiguration(
       sendNetworkInfo: true,
       printLogsToConsole: true,
@@ -59,6 +59,7 @@ Future<void> main() async {
         ? RumConfiguration(
             applicationId: applicationId,
             tracingSamplingRate: 100,
+            customEndpoint: customEndpoint,
           )
         : null,
   )..enableHttpTracking();